requirements.txt
# Основные зависимости для Telegram бота PROpitashka

# Telegram Bot Framework
aiogram==3.13.0

# База данных PostgreSQL
psycopg2-binary==2.9.9

# AI и генерация контента
google-generativeai==0.8.5
gigachat==0.1.35

# HTTP клиенты
aiohttp==3.9.3
requests==2.31.0

# Переводчик
translate==3.6.1

# Конфигурация
python-dotenv==1.0.1

# Утилиты
python-dateutil==2.9.0.post0

